Extract "Highlight Crime" action and reducer
