+++ frontend/package.json
@@ -12,6 +12,7 @@
     "react-leaflet": "^2.4.0",
     "react-redux": "^7.0.3",
     "react-scripts": "^3.0.1",
+    "reduce-reducers": "^1.0.4",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0"
