+++ frontend/src/index.tsx
@@ -1,26 +1,32 @@
+import CrimeMap from 'components/crimemap/crimeMap';
+import CrimeSelector from 'components/crimeselector/crimeSelector';
+import selectorReducer from 'components/crimeselector/crimeSelectorRedux';
+import CrimeViewer from 'components/crimeviewer/crimeViewer';
+import viewerReducer from 'components/crimeviewer/crimeViewerRedux';
+import { IStoreState } from 'components/types';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import reduceReducers from 'reduce-reducers';
 import { applyMiddleware, createStore } from 'redux';
 import { createLogger } from 'redux-logger';
 import thunkMiddleware from 'redux-thunk';
-import CrimeMap from 'components/crimemap/crimeMap';
-import CrimeSelector from 'components/crimeselector/crimeSelector';
-import CrimeViewer from 'components/crimeviewer/crimeViewer';
-import rootReducer from 'components/crimeselector/crimeSelectorRedux';
 import './index.css';
 import registerServiceWorker from './registerServiceWorker';
 
-const preloadedStore = {
+const preloadedStore: IStoreState= {
   allReportIds: [],
-  crimes: [],
+  crimes: [],  
+  day: '',
+  reportId: -1  
 }
 
+const rootReducer = reduceReducers(selectorReducer as any, viewerReducer as any)
 const loggerMiddleware = createLogger()
 
 const store = createStore(
   rootReducer,
-  preloadedStore,
+  preloadedStore as any,
   applyMiddleware(
     thunkMiddleware,
     loggerMiddleware
